# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for Epic Mail",
    "version": "1.0.0",
    "title": "Epic Mail"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "auth",
      "description": "Handles Signup and Signin"
    },
    {
      "name": "messages",
      "description": "Handles retreiving, sending and deleting mails"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a user account",
        "description": "",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string",
                  "example": "James"
                },
                "lastName": {
                  "type": "string",
                  "example": "Doe"
                },
                "userName": {
                  "type": "string",
                  "example": "JDoe"
                },
                "password": {
                  "type": "string",
                  "example": "secret"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "4567yjnjfn645"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing Parameter",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "Missing parameter"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login a user",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@test.com"
                },
                "password": {
                  "type": "string",
                  "example": "notell"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "4567yjnjfn645"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing Parameter",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "error": {
                  "type": "string",
                  "example": "Missing Parameter"
                }
              }
            }
          },
          "401": {
            "description": "Incorrect Credentials",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "error": {
                  "type": "string",
                  "example": "Missing Parameter"
                }
              }
            }
          }
        }
      }
    },
    "/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Fetch all received emails",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "createdOn": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Ipsum lorem"
                  },
                  "message": {
                    "type": "string",
                    "example": "Phasellus orci ante, euismod vitae quam sit amet, semper pulvinar risus"
                  },
                  "senderId": {
                    "type": "integer"
                  },
                  "receiverId": {
                    "type": "integer"
                  },
                  "parentMessageId": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Create or send an email.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "senderId": {
                  "type": "integer",
                  "example": 85839
                },
                "receiverId": {
                  "type": "integer",
                  "example": 2000
                },
                "subject": {
                  "type": "string",
                  "example": "Test Subject"
                },
                "message": {
                  "type": "string",
                  "example": "deido dsope gole aerople being maple trees derma terms"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "createdOn": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "subject": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "parentMessageId": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing Parameter",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "error": {
                  "type": "string",
                  "example": "Missing parameters"
                }
              }
            }
          }
        }
      }
    },
    "/messages/unread": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Fetch all unread received emails",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "createdOn": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Ipsum lorem"
                  },
                  "message": {
                    "type": "string",
                    "example": "Phasellus orci ante, euismod vitae quam sit amet, semper pulvinar risus"
                  },
                  "senderId": {
                    "type": "integer"
                  },
                  "receiverId": {
                    "type": "integer"
                  },
                  "parentMessageId": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "example": "unread"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/messages/sent": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Fetch sent emails",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "createdOn": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Ipsum lorem"
                  },
                  "message": {
                    "type": "string",
                    "example": "Phasellus orci ante, euismod vitae quam sit amet, semper pulvinar risus"
                  },
                  "senderId": {
                    "type": "integer"
                  },
                  "receiverId": {
                    "type": "integer"
                  },
                  "parentMessageId": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "example": "sent"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/messages/{messageId}": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Fetch a specific email record",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "ID of message to return",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "createdOn": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "subject": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "senderId": {
                      "type": "integer"
                    },
                    "receiverId": {
                      "type": "integer"
                    },
                    "parentMessageId": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Bad request. Message Id must be an Integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Message not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "messages"
        ],
        "summary": "Delete a specific email record",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "Message id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "createdOn": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "subject": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "senderId": {
                      "type": "integer"
                    },
                    "receiverId": {
                      "type": "integer"
                    },
                    "parentMessageId": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Bad request. Message Id must be an Integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Message not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}